+++ x/auction/auction.pb.go
@@ -4,18 +4,25 @@
 package auction
 
 import (
+	cosmossdk_io_math "cosmossdk.io/math"
 	fmt "fmt"
+	_ "github.com/cosmos/cosmos-proto"
+	types "github.com/cosmos/cosmos-sdk/types"
 	_ "github.com/cosmos/gogoproto/gogoproto"
 	proto "github.com/cosmos/gogoproto/proto"
+	github_com_cosmos_gogoproto_types "github.com/cosmos/gogoproto/types"
+	_ "google.golang.org/protobuf/types/known/timestamppb"
 	io "io"
 	math "math"
 	math_bits "math/bits"
+	time "time"
 )
 
 // Reference imports to suppress errors if they are not otherwise used.
 var _ = proto.Marshal
 var _ = fmt.Errorf
 var _ = math.Inf
+var _ = time.Kitchen
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the proto package it is being compiled against.
@@ -62,26 +69,120 @@ func (m *RewardsParams) XXX_DiscardUnknown() {
 
 var xxx_messageInfo_RewardsParams proto.InternalMessageInfo
 
+// Bid records a user bid
+type Bid struct {
+	Bidder []byte                `protobuf:"bytes,1,opt,name=bidder,proto3" json:"bidder,omitempty"`
+	Amount cosmossdk_io_math.Int `protobuf:"bytes,2,opt,name=amount,proto3,customtype=cosmossdk.io/math.Int" json:"amount"`
+}
+
+func (m *Bid) Reset()         { *m = Bid{} }
+func (m *Bid) String() string { return proto.CompactTextString(m) }
+func (*Bid) ProtoMessage()    {}
+func (*Bid) Descriptor() ([]byte, []int) {
+	return fileDescriptor_7a7eec280427e7e3, []int{1}
+}
+func (m *Bid) XXX_Unmarshal(b []byte) error {
+	return m.Unmarshal(b)
+}
+func (m *Bid) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
+	if deterministic {
+		return xxx_messageInfo_Bid.Marshal(b, m, deterministic)
+	} else {
+		b = b[:cap(b)]
+		n, err := m.MarshalToSizedBuffer(b)
+		if err != nil {
+			return nil, err
+		}
+		return b[:n], nil
+	}
+}
+func (m *Bid) XXX_Merge(src proto.Message) {
+	xxx_messageInfo_Bid.Merge(m, src)
+}
+func (m *Bid) XXX_Size() int {
+	return m.Size()
+}
+func (m *Bid) XXX_DiscardUnknown() {
+	xxx_messageInfo_Bid.DiscardUnknown(m)
+}
+
+var xxx_messageInfo_Bid proto.InternalMessageInfo
+
+// Auction Rewards
+type Rewards struct {
+	Rewards []types.Coin `protobuf:"bytes,1,rep,name=rewards,proto3" json:"rewards"`
+	EndsAt  time.Time    `protobuf:"bytes,2,opt,name=ends_at,json=endsAt,proto3,stdtime" json:"ends_at"`
+}
+
+func (m *Rewards) Reset()         { *m = Rewards{} }
+func (m *Rewards) String() string { return proto.CompactTextString(m) }
+func (*Rewards) ProtoMessage()    {}
+func (*Rewards) Descriptor() ([]byte, []int) {
+	return fileDescriptor_7a7eec280427e7e3, []int{2}
+}
+func (m *Rewards) XXX_Unmarshal(b []byte) error {
+	return m.Unmarshal(b)
+}
+func (m *Rewards) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
+	if deterministic {
+		return xxx_messageInfo_Rewards.Marshal(b, m, deterministic)
+	} else {
+		b = b[:cap(b)]
+		n, err := m.MarshalToSizedBuffer(b)
+		if err != nil {
+			return nil, err
+		}
+		return b[:n], nil
+	}
+}
+func (m *Rewards) XXX_Merge(src proto.Message) {
+	xxx_messageInfo_Rewards.Merge(m, src)
+}
+func (m *Rewards) XXX_Size() int {
+	return m.Size()
+}
+func (m *Rewards) XXX_DiscardUnknown() {
+	xxx_messageInfo_Rewards.DiscardUnknown(m)
+}
+
+var xxx_messageInfo_Rewards proto.InternalMessageInfo
+
 func init() {
 	proto.RegisterType((*RewardsParams)(nil), "umee.auction.v1.RewardsParams")
+	proto.RegisterType((*Bid)(nil), "umee.auction.v1.Bid")
+	proto.RegisterType((*Rewards)(nil), "umee.auction.v1.Rewards")
 }
 
 func init() { proto.RegisterFile("umee/auction/v1/auction.proto", fileDescriptor_7a7eec280427e7e3) }
 
 var fileDescriptor_7a7eec280427e7e3 = []byte{
-	// 184 bytes of a gzipped FileDescriptorProto
-	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0xe2, 0x92, 0x2d, 0xcd, 0x4d, 0x4d,
-	0xd5, 0x4f, 0x2c, 0x4d, 0x2e, 0xc9, 0xcc, 0xcf, 0xd3, 0x2f, 0x33, 0x84, 0x31, 0xf5, 0x0a, 0x8a,
-	0xf2, 0x4b, 0xf2, 0x85, 0xf8, 0x41, 0xd2, 0x7a, 0x30, 0xb1, 0x32, 0x43, 0x29, 0x91, 0xf4, 0xfc,
-	0xf4, 0x7c, 0xb0, 0x9c, 0x3e, 0x88, 0x05, 0x51, 0xa6, 0x64, 0xc4, 0xc5, 0x1b, 0x94, 0x5a, 0x9e,
-	0x58, 0x94, 0x52, 0x1c, 0x90, 0x58, 0x94, 0x98, 0x5b, 0x2c, 0xa4, 0xc8, 0xc5, 0x93, 0x94, 0x99,
-	0x12, 0x9f, 0x52, 0x5a, 0x94, 0x08, 0xd2, 0x29, 0xc1, 0xa8, 0xc0, 0xa8, 0xc1, 0x1c, 0xc4, 0x9d,
-	0x94, 0x99, 0xe2, 0x02, 0x15, 0x72, 0x72, 0x3f, 0xf1, 0x50, 0x8e, 0xe1, 0xc4, 0x23, 0x39, 0xc6,
-	0x0b, 0x8f, 0xe4, 0x18, 0x1f, 0x3c, 0x92, 0x63, 0x9c, 0xf0, 0x58, 0x8e, 0xe1, 0xc2, 0x63, 0x39,
-	0x86, 0x1b, 0x8f, 0xe5, 0x18, 0xa2, 0x34, 0xd3, 0x33, 0x4b, 0x32, 0x4a, 0x93, 0xf4, 0x92, 0xf3,
-	0x73, 0xf5, 0x41, 0x6e, 0xd0, 0xcd, 0x4b, 0x2d, 0x29, 0xcf, 0x2f, 0xca, 0x06, 0x73, 0xf4, 0xcb,
-	0xcc, 0xf4, 0x2b, 0x60, 0x2e, 0x4d, 0x62, 0x03, 0xbb, 0xc1, 0x18, 0x10, 0x00, 0x00, 0xff, 0xff,
-	0xa4, 0x42, 0x37, 0xb0, 0xcb, 0x00, 0x00, 0x00,
+	// 409 bytes of a gzipped FileDescriptorProto
+	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0x34, 0x52, 0xc1, 0x6e, 0x13, 0x31,
+	0x10, 0x5d, 0x13, 0x94, 0x80, 0x53, 0x84, 0xb4, 0x2a, 0x52, 0x1a, 0x09, 0x27, 0xe4, 0x14, 0x84,
+	0x62, 0x93, 0x20, 0x21, 0x71, 0xe0, 0xd0, 0x2d, 0x12, 0xea, 0x0d, 0x2d, 0x9c, 0xb8, 0x44, 0x76,
+	0x6c, 0xb6, 0x56, 0xb1, 0x5d, 0xd9, 0xde, 0x2d, 0x07, 0x6e, 0xfc, 0x40, 0x3f, 0xa6, 0x1f, 0x91,
+	0x63, 0xd5, 0x13, 0xe2, 0x50, 0x20, 0xf9, 0x91, 0xca, 0x6b, 0xef, 0xed, 0xcd, 0xbc, 0x79, 0x3b,
+	0x6f, 0xde, 0x1a, 0x3e, 0xaf, 0x95, 0x10, 0x84, 0xd6, 0x1b, 0x2f, 0x8d, 0x26, 0xcd, 0xb2, 0x83,
+	0xf8, 0xc2, 0x1a, 0x6f, 0xf2, 0xa7, 0x81, 0xc6, 0x5d, 0xaf, 0x59, 0x8e, 0x27, 0x95, 0x31, 0xd5,
+	0x77, 0x41, 0x5a, 0x9a, 0xd5, 0xdf, 0x88, 0x97, 0x4a, 0x38, 0x4f, 0xd5, 0x45, 0x54, 0x8c, 0x0f,
+	0x2b, 0x53, 0x99, 0x16, 0x92, 0x80, 0x52, 0xf7, 0x68, 0x63, 0x9c, 0x32, 0x6e, 0x1d, 0x89, 0x58,
+	0x24, 0x0a, 0xc5, 0x8a, 0x30, 0xea, 0x04, 0x69, 0x96, 0x4c, 0x78, 0xba, 0x24, 0x1b, 0x23, 0x93,
+	0x85, 0xd9, 0x0a, 0x3e, 0x29, 0xc5, 0x25, 0xb5, 0xdc, 0x7d, 0xa2, 0x96, 0x2a, 0x97, 0xbf, 0x80,
+	0x07, 0x4c, 0xf2, 0x35, 0xaf, 0x2d, 0x0d, 0xae, 0x46, 0x60, 0x0a, 0xe6, 0xbd, 0x72, 0xc8, 0x24,
+	0xff, 0x90, 0x5a, 0xb3, 0x9f, 0xb0, 0x57, 0x48, 0x9e, 0xbf, 0x86, 0x7d, 0x26, 0x39, 0x17, 0xb6,
+	0x9d, 0x39, 0x28, 0x46, 0xb7, 0xd7, 0x8b, 0xc3, 0xb4, 0xfc, 0x98, 0x73, 0x2b, 0x9c, 0xfb, 0xec,
+	0xad, 0xd4, 0x55, 0x99, 0xe6, 0xf2, 0x13, 0xd8, 0xa7, 0xca, 0xd4, 0xda, 0x8f, 0x1e, 0x4c, 0xc1,
+	0xfc, 0x71, 0xf1, 0x6a, 0x7b, 0x37, 0xc9, 0xfe, 0xdc, 0x4d, 0x9e, 0x45, 0x95, 0xe3, 0xe7, 0x58,
+	0x1a, 0xa2, 0xa8, 0x3f, 0xc3, 0xa7, 0xda, 0xdf, 0x5e, 0x2f, 0x60, 0xfa, 0xdc, 0xa9, 0xf6, 0x65,
+	0x92, 0xce, 0x7e, 0x01, 0x38, 0x48, 0x96, 0xf3, 0x77, 0x70, 0x60, 0x23, 0x1c, 0x81, 0x69, 0x6f,
+	0x3e, 0x5c, 0x1d, 0xe1, 0xa4, 0x08, 0xf7, 0xe2, 0x74, 0x2f, 0x3e, 0x31, 0x52, 0x17, 0x0f, 0xc3,
+	0xb2, 0xb2, 0x9b, 0xcf, 0xdf, 0xc3, 0x81, 0xd0, 0xdc, 0xad, 0x69, 0x34, 0x33, 0x5c, 0x8d, 0x71,
+	0x0c, 0x1f, 0x77, 0xe1, 0xe3, 0x2f, 0x5d, 0xf8, 0xc5, 0xa3, 0xa0, 0xbd, 0xfa, 0x3b, 0x01, 0x65,
+	0x3f, 0x88, 0x8e, 0x7d, 0xf1, 0x71, 0xfb, 0x1f, 0x65, 0xdb, 0x1d, 0x02, 0x37, 0x3b, 0x04, 0xfe,
+	0xed, 0x10, 0xb8, 0xda, 0xa3, 0xec, 0x66, 0x8f, 0xb2, 0xdf, 0x7b, 0x94, 0x7d, 0x7d, 0x59, 0x49,
+	0x7f, 0x56, 0x33, 0xbc, 0x31, 0x8a, 0x84, 0x7f, 0xbc, 0xd0, 0xc2, 0x5f, 0x1a, 0x7b, 0xde, 0x16,
+	0xa4, 0x79, 0x4b, 0x7e, 0x74, 0x2f, 0x81, 0xf5, 0xdb, 0x75, 0x6f, 0xee, 0x03, 0x00, 0x00, 0xff,
+	0xff, 0xa0, 0x8c, 0xc4, 0xb7, 0x2b, 0x02, 0x00, 0x00,
 }
 
 func (m *RewardsParams) Marshal() (dAtA []byte, err error) {
@@ -112,6 +213,91 @@ func (m *RewardsParams) MarshalToSizedBuffer(dAtA []byte) (int, error) {
 	return len(dAtA) - i, nil
 }
 
+func (m *Bid) Marshal() (dAtA []byte, err error) {
+	size := m.Size()
+	dAtA = make([]byte, size)
+	n, err := m.MarshalToSizedBuffer(dAtA[:size])
+	if err != nil {
+		return nil, err
+	}
+	return dAtA[:n], nil
+}
+
+func (m *Bid) MarshalTo(dAtA []byte) (int, error) {
+	size := m.Size()
+	return m.MarshalToSizedBuffer(dAtA[:size])
+}
+
+func (m *Bid) MarshalToSizedBuffer(dAtA []byte) (int, error) {
+	i := len(dAtA)
+	_ = i
+	var l int
+	_ = l
+	{
+		size := m.Amount.Size()
+		i -= size
+		if _, err := m.Amount.MarshalTo(dAtA[i:]); err != nil {
+			return 0, err
+		}
+		i = encodeVarintAuction(dAtA, i, uint64(size))
+	}
+	i--
+	dAtA[i] = 0x12
+	if len(m.Bidder) > 0 {
+		i -= len(m.Bidder)
+		copy(dAtA[i:], m.Bidder)
+		i = encodeVarintAuction(dAtA, i, uint64(len(m.Bidder)))
+		i--
+		dAtA[i] = 0xa
+	}
+	return len(dAtA) - i, nil
+}
+
+func (m *Rewards) Marshal() (dAtA []byte, err error) {
+	size := m.Size()
+	dAtA = make([]byte, size)
+	n, err := m.MarshalToSizedBuffer(dAtA[:size])
+	if err != nil {
+		return nil, err
+	}
+	return dAtA[:n], nil
+}
+
+func (m *Rewards) MarshalTo(dAtA []byte) (int, error) {
+	size := m.Size()
+	return m.MarshalToSizedBuffer(dAtA[:size])
+}
+
+func (m *Rewards) MarshalToSizedBuffer(dAtA []byte) (int, error) {
+	i := len(dAtA)
+	_ = i
+	var l int
+	_ = l
+	n1, err1 := github_com_cosmos_gogoproto_types.StdTimeMarshalTo(m.EndsAt, dAtA[i-github_com_cosmos_gogoproto_types.SizeOfStdTime(m.EndsAt):])
+	if err1 != nil {
+		return 0, err1
+	}
+	i -= n1
+	i = encodeVarintAuction(dAtA, i, uint64(n1))
+	i--
+	dAtA[i] = 0x12
+	if len(m.Rewards) > 0 {
+		for iNdEx := len(m.Rewards) - 1; iNdEx >= 0; iNdEx-- {
+			{
+				size, err := m.Rewards[iNdEx].MarshalToSizedBuffer(dAtA[:i])
+				if err != nil {
+					return 0, err
+				}
+				i -= size
+				i = encodeVarintAuction(dAtA, i, uint64(size))
+			}
+			i--
+			dAtA[i] = 0xa
+		}
+	}
+	return len(dAtA) - i, nil
+}
+
 func encodeVarintAuction(dAtA []byte, offset int, v uint64) int {
 	offset -= sovAuction(v)
 	base := offset
@@ -132,6 +318,38 @@ func (m *RewardsParams) Size() (n int) {
 	if m.BidDuration != 0 {
 		n += 1 + sovAuction(uint64(m.BidDuration))
 	}
+	return n
+}
+
+func (m *Bid) Size() (n int) {
+	if m == nil {
+		return 0
+	}
+	var l int
+	_ = l
+	l = len(m.Bidder)
+	if l > 0 {
+		n += 1 + l + sovAuction(uint64(l))
+	}
+	l = m.Amount.Size()
+	n += 1 + l + sovAuction(uint64(l))
+	return n
+}
+
+func (m *Rewards) Size() (n int) {
+	if m == nil {
+		return 0
+	}
+	var l int
+	_ = l
+	if len(m.Rewards) > 0 {
+		for _, e := range m.Rewards {
+			l = e.Size()
+			n += 1 + l + sovAuction(uint64(l))
+		}
+	}
+	l = github_com_cosmos_gogoproto_types.SizeOfStdTime(m.EndsAt)
+	n += 1 + l + sovAuction(uint64(l))
 	return n
 }
 
@@ -189,6 +407,241 @@ func (m *RewardsParams) Unmarshal(dAtA []byte) error {
 					break
 				}
 			}
+		default:
+			iNdEx = preIndex
+			skippy, err := skipAuction(dAtA[iNdEx:])
+			if err != nil {
+				return err
+			}
+			if (skippy < 0) || (iNdEx+skippy) < 0 {
+				return ErrInvalidLengthAuction
+			}
+			if (iNdEx + skippy) > l {
+				return io.ErrUnexpectedEOF
+			}
+			iNdEx += skippy
+		}
+	}
+
+	if iNdEx > l {
+		return io.ErrUnexpectedEOF
+	}
+	return nil
+}
+func (m *Bid) Unmarshal(dAtA []byte) error {
+	l := len(dAtA)
+	iNdEx := 0
+	for iNdEx < l {
+		preIndex := iNdEx
+		var wire uint64
+		for shift := uint(0); ; shift += 7 {
+			if shift >= 64 {
+				return ErrIntOverflowAuction
+			}
+			if iNdEx >= l {
+				return io.ErrUnexpectedEOF
+			}
+			b := dAtA[iNdEx]
+			iNdEx++
+			wire |= uint64(b&0x7F) << shift
+			if b < 0x80 {
+				break
+			}
+		}
+		fieldNum := int32(wire >> 3)
+		wireType := int(wire & 0x7)
+		if wireType == 4 {
+			return fmt.Errorf("proto: Bid: wiretype end group for non-group")
+		}
+		if fieldNum <= 0 {
+			return fmt.Errorf("proto: Bid: illegal tag %d (wire type %d)", fieldNum, wire)
+		}
+		switch fieldNum {
+		case 1:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field Bidder", wireType)
+			}
+			var byteLen int
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowAuction
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				byteLen |= int(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			if byteLen < 0 {
+				return ErrInvalidLengthAuction
+			}
+			postIndex := iNdEx + byteLen
+			if postIndex < 0 {
+				return ErrInvalidLengthAuction
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			m.Bidder = append(m.Bidder[:0], dAtA[iNdEx:postIndex]...)
+			if m.Bidder == nil {
+				m.Bidder = []byte{}
+			}
+			iNdEx = postIndex
+		case 2:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field Amount", wireType)
+			}
+			var stringLen uint64
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowAuction
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				stringLen |= uint64(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			intStringLen := int(stringLen)
+			if intStringLen < 0 {
+				return ErrInvalidLengthAuction
+			}
+			postIndex := iNdEx + intStringLen
+			if postIndex < 0 {
+				return ErrInvalidLengthAuction
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			if err := m.Amount.Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
+				return err
+			}
+			iNdEx = postIndex
+		default:
+			iNdEx = preIndex
+			skippy, err := skipAuction(dAtA[iNdEx:])
+			if err != nil {
+				return err
+			}
+			if (skippy < 0) || (iNdEx+skippy) < 0 {
+				return ErrInvalidLengthAuction
+			}
+			if (iNdEx + skippy) > l {
+				return io.ErrUnexpectedEOF
+			}
+			iNdEx += skippy
+		}
+	}
+
+	if iNdEx > l {
+		return io.ErrUnexpectedEOF
+	}
+	return nil
+}
+func (m *Rewards) Unmarshal(dAtA []byte) error {
+	l := len(dAtA)
+	iNdEx := 0
+	for iNdEx < l {
+		preIndex := iNdEx
+		var wire uint64
+		for shift := uint(0); ; shift += 7 {
+			if shift >= 64 {
+				return ErrIntOverflowAuction
+			}
+			if iNdEx >= l {
+				return io.ErrUnexpectedEOF
+			}
+			b := dAtA[iNdEx]
+			iNdEx++
+			wire |= uint64(b&0x7F) << shift
+			if b < 0x80 {
+				break
+			}
+		}
+		fieldNum := int32(wire >> 3)
+		wireType := int(wire & 0x7)
+		if wireType == 4 {
+			return fmt.Errorf("proto: Rewards: wiretype end group for non-group")
+		}
+		if fieldNum <= 0 {
+			return fmt.Errorf("proto: Rewards: illegal tag %d (wire type %d)", fieldNum, wire)
+		}
+		switch fieldNum {
+		case 1:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field Rewards", wireType)
+			}
+			var msglen int
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowAuction
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				msglen |= int(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			if msglen < 0 {
+				return ErrInvalidLengthAuction
+			}
+			postIndex := iNdEx + msglen
+			if postIndex < 0 {
+				return ErrInvalidLengthAuction
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			m.Rewards = append(m.Rewards, types.Coin{})
+			if err := m.Rewards[len(m.Rewards)-1].Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
+				return err
+			}
+			iNdEx = postIndex
+		case 2:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field EndsAt", wireType)
+			}
+			var msglen int
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowAuction
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				msglen |= int(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			if msglen < 0 {
+				return ErrInvalidLengthAuction
+			}
+			postIndex := iNdEx + msglen
+			if postIndex < 0 {
+				return ErrInvalidLengthAuction
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			if err := github_com_cosmos_gogoproto_types.StdTimeUnmarshal(&m.EndsAt, dAtA[iNdEx:postIndex]); err != nil {
+				return err
+			}
+			iNdEx = postIndex
 		default:
 			iNdEx = preIndex
 			skippy, err := skipAuction(dAtA[iNdEx:])