+++ CHANGELOG.md
@@ -49,11 +49,14 @@ Ref: https://keepachangelog.com/en/1.0.0/
 ### Features
 
 - [2472](https://github.com/umee-network/umee/pull/2472) un-wire the `crisis` module from umee app.
+- [2500](https://github.com/umee-network/umee/pull/2500) (x/leverage): add Rewards Auction fees and `params.rewards_auction_factor`.
 
 ### Improvements
 
 - [2474](https://github.com/umee-network/umee/pull/2474) (proto) add `gogo.messagename_all` option to all messages.
 - [2494](https://github.com/umee-network/umee/pull/2494) Use go 1.22
+- [2495](https://github.com/umee-network/umee/pull/2495) (util) `KeyWithUint64` and `KeyWithUint32` helper functions for db keys.
+- [2504](https://github.com/umee-network/umee/pull/2504) bump `wasmd` to v0.45.0.
 
 ### Bug Fixes